pyproject.toml
[project]
name = "dw_audit_bot"
version = "0.0.1"
dependencies = [
    "hikari",
    "hikari-crescent",
    "python-dotenv",
    "pandas",
    "uvloop",
]

[project.scripts]
dw_audit_bot = "bot:main"

[tool.setuptools.packages]
find = {}

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 120
